// frontend/src/api/models/analysisAuditTransaction.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Splice API
 * Financial account management API
 * OpenAPI spec version: 1.0
 */
import type { AnalysisAuditTransactionAmount } from './analysisAuditTransactionAmount'

export interface AnalysisAuditTransaction {
  id: string
  activityDate: string
  /** @nullable */
  merchantName: string | null
  /** @nullable */
  originalDescription: string | null
  accountName: string
  categoryPrimary: string
  /** @nullable */
  categoryDetailed: string | null
  amount: AnalysisAuditTransactionAmount
}