// frontend/src/api/models/transactionAnalysisAuditResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Splice API
 * Financial account management API
 * OpenAPI spec version: 1.0
 */
import type { TransactionAnalysisAuditResponseRowsItem } from './transactionAnalysisAuditResponseRowsItem'

export interface TransactionAnalysisAuditResponse {
  startDate: string
  endDate: string
  neutralizationLookaroundDays: number
  rows: TransactionAnalysisAuditResponseRowsItem[]
}